Split category objects
