+++ packages/common/src/shared/categoryList.ts
@@ -1,0 +1,4 @@
+import { CategoryType } from 'types';
+import { categoryQueryString } from './categoryQueryString';
+
+export const categoryList = Object.keys(categoryQueryString) as CategoryType[];
