+++ packages/common/src/shared/categoryQueryString.ts
@@ -1,0 +1,11 @@
+import { CategoryQueryStringType, CategoryType } from 'types';
+
+type CategoryQueryString = {
+  [key in CategoryType]: CategoryQueryStringType;
+};
+
+export const categoryQueryString: CategoryQueryString = {
+  notice: 'NOTICE',
+  newsletter: 'FAMILY_NEWSLETTER',
+  gallery: 'EVENT_GALLERY',
+} as const;
